refactor: handleFormSubmission to controller
