+++ src/controllers/submitForm.ts
@@ -1,12 +1,11 @@
-const submitForm = {};
-const url = '/test';
+const url = "/test";
 
 export async function POST(url, body) {
   try {
     const res = await fetch(url, {
-      method: 'POST',
+      method: "POST",
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
       body,
     });
@@ -15,4 +14,19 @@ export async function POST(url, body) {
   } catch (err) {
     console.log(err);
   }
+}
+
+export function handleFormSubmission() {
+  const form = document.querySelector("form");
+  const formData = new FormData(form);
+  const testParam = new URLSearchParams();
+  const testBody = {};
+  for (const [key, value] of formData) {
+    if (key === "htmlTag" && typeof value === "string") {
+      testParam.append(key, value);
+    } else {
+      testBody[key] = value;
+    }
+  }
+  let response = POST("/test?" + testParam, JSON.stringify(testBody));
 }
